Share prompt URL sanitizer
